+++ src/components/manage/Tiles/Title/Edit.jsx
@@ -1,0 +1,93 @@
+/**
+ * Edit title tile.
+ * @module components/manage/Tiles/Title/Edit
+ */
+
+import React, { Component } from 'react';
+import { Map } from 'immutable';
+import PropTypes from 'prop-types';
+import { stateFromHTML } from 'draft-js-import-html';
+import { Editor, DefaultDraftBlockRenderMap, EditorState } from 'draft-js';
+
+const blockRenderMap = Map({
+  unstyled: {
+    element: 'h1',
+  },
+});
+
+const extendedBlockRenderMap = DefaultDraftBlockRenderMap.merge(blockRenderMap);
+
+/**
+ * Edit title tile class.
+ * @class Edit
+ * @extends Component
+ */
+export default class Edit extends Component {
+  /**
+   * Property types.
+   * @property {Object} propTypes Property types.
+   * @static
+   */
+  static propTypes = {
+    properties: PropTypes.objectOf(PropTypes.any).isRequired,
+    onChange: PropTypes.func.isRequired,
+  };
+
+  /**
+   * Constructor
+   * @method constructor
+   * @param {Object} props Component properties
+   * @constructs WysiwygEditor
+   */
+  constructor(props) {
+    super(props);
+
+    if (!__SERVER__) {
+      let editorState;
+      if (props.properties && props.properties.title) {
+        const contentState = stateFromHTML(props.properties.title);
+        editorState = EditorState.createWithContent(contentState);
+      } else {
+        editorState = EditorState.createEmpty();
+      }
+      this.state = { editorState };
+    }
+
+    this.onChange = this.onChange.bind(this);
+  }
+
+  /**
+   * Change handler
+   * @method onChange
+   * @param {object} editorState Editor state.
+   * @returns {undefined}
+   */
+  onChange(editorState) {
+    this.setState({ editorState });
+    this.props.onChange({
+      properties: {
+        title: editorState.getCurrentContent().getPlainText(),
+      },
+    });
+  }
+
+  /**
+   * Render method.
+   * @method render
+   * @returns {string} Markup for the component.
+   */
+  render() {
+    if (__SERVER__) {
+      return <div />;
+    }
+    return (
+      <Editor
+        onChange={this.onChange}
+        editorState={this.state.editorState}
+        blockRenderMap={extendedBlockRenderMap}
+        handleReturn={() => true}
+        blockStyleFn={() => 'documentFirstHeading'}
+      />
+    );
+  }
+}